Make search depth a u8

+++ crates/grob_search/src/transposition.rs
@@ -19,7 +19,7 @@ pub struct Transposition {
     /// The [`Score`] of the position on the basis of the past search.
     pub score: Score,
     /// The depth of the past search.
-    pub depth: u64, // TODO: is u64 too large?
+    pub depth: u8, // TODO: is u64 too large?
 }
 
 /// A [transposition table].
